% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{auto.page}
\alias{auto.page}
\title{Automatically roll-over a search when paginated}
\usage{
auto.page(f, max.pages = 10)
}
\arguments{
\item{f}{call to execute}

\item{max.pages}{max number of pages to traverse (exits earlier if fail or complete)}
}
\value{
evaluates the call and returns the content

\feedbackfooter{'jonocarroll/githubtools'}{'R/scan.R'}{TRUE}
}
\description{
Automatically roll-over a search when paginated
}

