% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{build_html}
\alias{build_html}
\title{Generate a HTML Page of GitHub Tile Charts}
\usage{
build_html(pkg = NULL, img.dir = ".", viewer.pane = FALSE)
}
\arguments{
\item{pkg}{packages to include in the page}

\item{img.dir}{where the images are stored}

\item{viewer.pane}{(unused) logical. Use the inbuilt viewer pane?}
}
\value{
location of the stored .html file

\feedbackfooter{'jonocarroll/githubtools'}{'R/scan.R'}{TRUE}
}
\description{
Generate a HTML Page of GitHub Tile Charts
}

