% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{check_all_github}
\alias{check_all_github}
\title{Obtain commit stats for one or more GitHub repos}
\usage{
check_all_github(pkg = NULL, img.dir = ".")
}
\arguments{
\item{img.dir}{}
}
\value{

}
\description{
Obtain commit stats for one or more GitHub repos
}

