% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_prep.R
\name{create_github_chart}
\alias{create_github_chart}
\title{Create a GitHub-style Tile Chart with Social Network Color Palette}
\usage{
create_github_chart(gh_data, user, network = c("GitHub", "Twitter",
  "StackOverflow"))
}
\arguments{
\item{gh_data}{data prepared with \link{\code{prepare_for_github_chart}}}

\item{user}{user ID to add to the subtitle}

\item{network}{which color palette to use, styled after a network (GitHub, Twitter, or StackOverflow). 
Case insensitive, but will be also be used in the title.}
}
\value{
a ggplot2 object for printing
}
\description{
Create a GitHub-style Tile Chart with Social Network Color Palette
}

