% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{github_setup}
\alias{github_setup}
\title{Perform GitHub OAuth setup.}
\usage{
github_setup()
}
\value{
a github context object that is used in every github API call issued by the github package.
}
\description{
#@param config_file location of the configuration file, default ~/.githuburlcheckr
}
\details{
#@param echo if TRUE print the credentials read from the file to the console.
}
\examples{
\dontrun{
ctx <- github_setup()
}
}

