% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_prep.R
\name{prepare_for_github_chart}
\alias{prepare_for_github_chart}
\title{Prepare Daily Data Ready for Use in a GitHub-Style Chart}
\usage{
prepare_for_github_chart(data_agg, primaryData, secondaryData = "dummy")
}
\arguments{
\item{data_agg}{aggregated daily data}

\item{primaryData}{name of the primary data column to plot daily}

\item{secondaryData}{name of the secondary data column to label each daily tile (optional).}
}
\value{
a list object ready for \link{\code{create_github_chart}}.
}
\description{
Prepare Daily Data Ready for Use in a GitHub-Style Chart
}

