% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_github.R
\name{scale_fill_social}
\alias{scale_fill_social}
\alias{social_pal}
\title{Social networking color palettes (discrete)}
\usage{
scale_fill_social(network = c("GitHub", "Twitter", "StackOverflow"), ...)

social_pal(network = c("GitHub", "Twitter", "StackOverflow"))
}
\arguments{
\item{network}{which palette to use}
}
\description{
The hues in each palette are:
\itemize{
\item GitHub greens
\item Twitter blues
\item StackOverflow oranges
}
}
\examples{
library(scales)
show_col(social_pal("GitHub")(6))
show_col(social_pal("Twitter")(6))
show_col(social_pal("StackOverflow")(6))
}

