% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{scan_gh_pkgs}
\alias{scan_gh_pkgs}
\title{Scan and analyse GitHub R packages}
\usage{
scan_gh_pkgs(pkg = NULL, img.dir = ".", max.commits = 200)
}
\arguments{
\item{pkg}{package to check (local or external)}

\item{img.dir}{where to store generated images}
}
\value{
NULL (used for the side effect of generating a .html file in \code{img.dir})

\feedbackfooter{'jonocarroll/githubtools'}{'R/scan.R'}{TRUE}
}
\description{
Scan and analyse GitHub R packages
}

