% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_github.R
\name{theme_github}
\alias{theme_github}
\title{ggplot2 theme based on the GitHub contribution tile chart}
\usage{
theme_github()
}
\value{
An object of class \code{\link{theme}}.
}
\description{
Style a tile chart similar to those on \emph{GitHub} profiles.
}
\details{
\code{theme_github} implements the standard green tile chart as 
seen on GitHub profiles. Alternatively, uses a different color scheme
in the style of another popular site, e.g. twitter, stackoverflow.
}
\examples{
library("ggplot2")
## get data
p <- ggplot(commit_data) + geom_tile(aes(x=c.week, y=c.day)) + theme_github()

## Twitter colors
p + theme_github(alt="Twitter")

}

