% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan.R
\name{view_all_sources}
\alias{view_all_sources}
\title{View Data Related to All Installed Packages}
\usage{
view_all_sources(github.only = FALSE)
}
\value{
data.frame of installed package information

\feedbackfooter{'jonocarroll/githubtools'}{'R/scan.R'}{TRUE}
}
\description{
This may return a large \code{data.frame} depending on how many packages
you have installed (not just loaded).
}

